+++ packages/ui/src/lib/accordion/accordion-subtitle.ts
@@ -1,0 +1,8 @@
+import { Directive, inject, TemplateRef } from '@angular/core';
+
+@Directive({
+  selector: 'ng-template[elAccordionSubtitle]',
+})
+export class ElAccordionSubtitle {
+  readonly template = inject(TemplateRef<unknown>);
+}