+++ packages/ui/src/lib/accordion/accordion-title.ts
@@ -1,0 +1,8 @@
+import { Directive, inject, TemplateRef } from '@angular/core';
+
+@Directive({
+  selector: 'ng-template[elAccordionTitle]',
+})
+export class ElAccordionTitle {
+  readonly template = inject(TemplateRef<unknown>);
+}